Walking stick is now "shield". New, longer model.
A shield was a better fit for what a walking stick is, because:

* Shields don't sheath when you put away your weapons (and the running
animation looks really good with the walking stick in a shield slot now)
* You can't dual-wield shields
* It was always meant as more of an accessory than a weapon

It is now a Light Armor shield with 10 armor. It now grants a Fortify
Stamina bonus (+10 Stamina) instead of a speed bonus, which was always
almost unnoticeable and prone to breaking / complicated to implement.

The new model is 20% longer and 20% thinner, looking decidedly less
"log-like" and more appropriate.

The world object now makes wooden sounds and rolls around appropriately.